decoder: fix i32 overflow in dequantise when Pq=1 (16-bit quantiser)
A local 60 s fuzz session on the `decode` cargo-fuzz target uncovered a
`render_from_coefs` panic on a `Pq = 1` (16-bit quantiser, T.81
§B.2.4.1) input: the dequantise multiplication `block[k] * qt.values[k]
as i32` overflowed `i32` when the quantiser sat at 65535 and the
coefficient was near the top of the DCT range. The IDCT input is `f32`
either way, so the cleanest fix is to perform the multiplication in
`f32` directly — its 24-bit mantissa carries the product without
overflow. Three call sites — baseline 8-bit, 12-bit, and progressive —
all received the same one-line correction.

Fuzz-found crash artifact `crash-ee0cdd45…` now replays in 0 ms.
Re-running the 60 s `decode` baseline against the patched binary
produced 25 694 runs at cov 2023 / ft 7670 with zero crashes; the four
prior pre-existing crash artifacts in `fuzz/artifacts/decode/` also
replay clean (they were recorded earlier in bring-up and previously
fixed). All 144 in-crate tests pass.

+++ src/decoder.rs
@@ -788,7 +788,10 @@ fn decode_scan(
                         let mut fblock = [0.0f32; 64];
                         for k in 0..64 {
                             // `block` is natural-order; `qt.values` is natural-order.
-                            fblock[k] = (block[k] * qt.values[k] as i32) as f32;
+                            // Multiply in f32 to avoid i32 overflow when Pq=1 (16-bit
+                            // quantiser, values up to 65535) meets a coefficient at the
+                            // top of its range — the IDCT input is f32 either way.
+                            fblock[k] = block[k] as f32 * qt.values[k] as f32;
                         }
                         idct8x8(&mut fblock);
                         // Write 8×8 to component buffer at position (mx * H + bx, my * V + by).
@@ -1781,7 +1784,10 @@ fn render_from_coefs(
                 let block = &coefs[ci][by * blocks_x + bx];
                 let mut fblock = [0.0f32; 64];
                 for k in 0..64 {
-                    fblock[k] = (block[k] * qt.values[k] as i32) as f32;
+                    // Multiply in f32 to avoid i32 overflow when Pq=1 (16-bit
+                    // quantiser, values up to 65535) meets a coefficient at the
+                    // top of its range — the IDCT input is f32 either way.
+                    fblock[k] = block[k] as f32 * qt.values[k] as f32;
                 }
                 idct8x8(&mut fblock);
                 let dst_x0 = bx * 8;
@@ -2022,7 +2028,10 @@ fn render_from_coefs_12bit(
                 let block = &coefs[ci][by * blocks_x + bx];
                 let mut fblock = [0.0f32; 64];
                 for k in 0..64 {
-                    fblock[k] = (block[k] * qt.values[k] as i32) as f32;
+                    // Multiply in f32 to avoid i32 overflow when Pq=1 (16-bit
+                    // quantiser, values up to 65535) meets a coefficient at the
+                    // top of its range — the IDCT input is f32 either way.
+                    fblock[k] = block[k] as f32 * qt.values[k] as f32;
                 }
                 idct8x8(&mut fblock);
                 let dst_x0 = bx * 8;
